backend/requirements.txt
click==7.1.2
config==0.5.0.post0
Flask==1.1.2
Flask-Cors==3.0.10
flask-marshmallow==0.14.0
Flask-SQLAlchemy==2.5.1
greenlet==1.0.0
importlib-metadata==4.0.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
marshmallow==3.11.1
marshmallow-sqlalchemy==0.24.2
mysqlclient==2.0.3
PyJWT==1.7.1
six==1.15.0
SQLAlchemy==1.4.9
typing-extensions==3.7.4.3
Werkzeug==1.0.1
zipp==3.4.1
